https://github.com/javaConductor/mash/issues/8 moved some things around
