+++ src/processors/aggregateExtraction.js
